requirements.txt
python-dotenv
os
google-generativeai
streamlit